자신의 정보만 수정 가능

+++ backend/src/main/java/com/example/backend/controller/member/MemberController.java
@@ -4,6 +4,7 @@ import com.example.backend.dto.member.Member;
 import com.example.backend.dto.member.MemberEdit;
 import com.example.backend.service.member.MemberService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Map;
 
+@Slf4j
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/member")
@@ -33,69 +35,74 @@ public class MemberController {
             // 로그인 실패
             return ResponseEntity.status(401)
                     .body(Map.of("message", Map.of("type", "warning",
-                            "text", "로그인을 실패했습니다.")));
+                            "text", "아이디와 암호를 확인해주세요.")));
         }
     }
 
     @PutMapping("update")
-    public ResponseEntity<Map<String, Map<String, String>>> update(@RequestBody MemberEdit member) {
-        if (service.update(member)) {
-            //잘됨
-            return ResponseEntity.ok(Map.of("message",
-                    Map.of("type", "success",
-                            "text", "회원정보를 수정하였습니다.")));
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Map<String, Object>> update(
+            @RequestBody MemberEdit member,
+            Authentication authentication) {
+        if (service.hasAccess(member.getId(), authentication)) {
+
+            if (service.update(member)) {
+                //잘됨
+                return ResponseEntity.ok(Map.of("message",
+                        Map.of("type", "success",
+                                "text", "회원정보를 수정하였습니다.")));
+            } else {
+                return ResponseEntity.badRequest()
+                        .body(Map.of("message",
+                                Map.of("type", "warning",
+                                        "text", "정확한 정보를 입력해주세요.")));
+            }
         } else {
-            return ResponseEntity.badRequest()
-                    .body(Map.of("message",
-                            Map.of("type", "warning",
-                                    "text", "정확한 정보를 입력해주세요.")));
+            return ResponseEntity.status(403).build();
         }
+
+
     }
 
     @DeleteMapping("remove")
-    public ResponseEntity<Map<String, Object>> remove(@RequestBody Member member) {
-        if (service.remove(member)) {
-            //잘됨
-            return ResponseEntity.ok(Map.of("message",
-                    Map.of("type", "success",
-                            "text", "회원정보를 삭제하였습니다.")));
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Map<String, Object>> remove(
+            @RequestBody Member member,
+            Authentication authentication) {
+        if (service.hasAccess(member.getId(), authentication)) {
+
+            if (service.remove(member)) {
+                //잘됨
+                return ResponseEntity.ok(Map.of("message",
+                        Map.of("type", "success",
+                                "text", "회원정보를 삭제하였습니다.")));
+            } else {
+                return ResponseEntity.badRequest()
+                        .body(Map.of("message",
+                                Map.of("type", "warning",
+                                        "text", "정확한 정보를 입력해주세요.")));
+            }
         } else {
-            return ResponseEntity.badRequest()
-                    .body(Map.of("message",
-                            Map.of("type", "warning",
-                                    "text", "정확한 정보를 입력해주세요.")));
+            return ResponseEntity.status(403).build();
         }
     }
 
     @GetMapping("{id}")
     @PreAuthorize("isAuthenticated()")
     public ResponseEntity<Member> getMember(@PathVariable String id, Authentication auth) {
+
         if (service.hasAccess(id, auth)) {
             return ResponseEntity.ok(service.get(id));
         } else {
             return ResponseEntity.status(403).build();
         }
+
     }
+
 
     @GetMapping("list")
     public List<Member> list() {
         return service.list();
-    }
-
-    @GetMapping(value = "check", params = "id")
-    public ResponseEntity<Map<String, Object>> checkId(@RequestParam String id) {
-        if (service.checkId(id)) {
-            // 이미 있으면
-            return ResponseEntity.ok().body(Map.of(
-                    "message", Map.of("type", "warning", "text", "이미 사용중인 아이디입니다."),
-                    "available", false));
-
-        } else {
-            // 없으면
-            return ResponseEntity.ok().body(Map.of("message",
-                    Map.of("type", "success", "text", "사용 가능한 아이디입니다."),
-                    "available", true));
-        }
     }
 
     @GetMapping(value = "check", params = "email")
@@ -114,22 +121,37 @@ public class MemberController {
         }
     }
 
+    @GetMapping(value = "check", params = "id")
+    public ResponseEntity<Map<String, Object>> checkId(@RequestParam String id) {
+        if (service.checkId(id)) {
+            // 이미 있으면
+            return ResponseEntity.ok().body(Map.of(
+                    "message", Map.of("type", "warning", "text", "이미 사용중인 아이디 입니다."),
+                    "available", false)
+            );
+        } else {
+            // 없으면
+            return ResponseEntity.ok().body(Map.of(
+                    "message", Map.of("type", "info", "text", "사용 가능한 아이디 입니다."),
+                    "available", true));
+        }
+    }
+
     @PostMapping("signup")
     public ResponseEntity<Map<String, Object>> signup(@RequestBody Member member) {
         try {
 
             if (service.add(member)) {
                 return ResponseEntity.ok().body(Map.of("message",
-                        Map.of("type", "success", "text", "회원가입되었습니다.")));
-
+                        Map.of("type", "success", "text", "회원 가입되었습니다.")));
             } else {
                 return ResponseEntity.internalServerError().body(Map.of("message",
-                        Map.of("type", "error", "text", "회원가입 중 문제가 발생하였습니다.")));
-
+                        Map.of("type", "error", "text", "회원 가입 중 문제가 발생하였습니다.")));
             }
         } catch (DuplicateKeyException e) {
             return ResponseEntity.internalServerError().body(Map.of("message",
                     Map.of("type", "error", "text", "이미 존재하는 아이디입니다.")));
         }
     }
+
 }
